Third implementation of the 2by4 decoder
This is reimplementation of the 2 to 4 line decoder with 2 more submodules. Therefore 2 new classes anding_gate and inverter are introduced to model the above functionality.
The 1 to 2 line decoder is implemented using 2 NOT gates and 2 AND gates as shown in the moc image below. The 1 binary input is fed into the port x1. The default binary input port is held high at all times and according to the binary input value fed into port x1, the truth table shown below is generated.
For simplicity, the default binary input is labelled as d_x in the truth table.