Creation of four bit shift register using the DFFs
The detailed Dflipflop folder was used and all files retained. Created four_bitshiftreg .cc and .h files. Implemented four bit shift register in the function in the four_bitshiftreg.cc file and called it in the .h file. A register instance was created and connected to the driver and monitor in the main method file. the Makefile was edited to accomodate the new .cc file to be compiled to a .o file.